backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
channels>=4.1
channels-redis>=4.2
psycopg2-binary>=2.9
asgiref>=3.7
daphne>=4.1
websocket-client>=1.7
